requirements.txt
anthropic>=0.84.0
exa-py>=1.0.0
linkup-sdk>=0.12.0
requests>=2.31.0
python-dotenv>=1.0.0
